Made PointStream a singletone. Fixed the DisplayFreq and DisplaySInglePoint to work with PointStream. Made an Error view for incase an exception occurs.

+++ Ex3/Ex3/Models/Point.cs
@@ -11,7 +11,7 @@ namespace Ex3.Models
         public double X { get; set; }
         public double Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             this.X = x;
             this.Y = y;
